+++ CHANGELOG.md
@@ -37,6 +37,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## Unreleased
 
+## v0.14.1
+
 ### API Breaking
 
 * [#228](https://github.com/babylonlabs-io/babylon/pull/228) Add inclusion height to early unbonding event